Pushed more changes based on feedback from PR.
