+++ src/dotnet/CarbonAware.Tools.AWSRegionTestDataGenerator/CarbonAware.Tools.AWSRegionTestDataGenerator.csproj
@@ -18,7 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
 
   <ItemGroup>
