+++ src/dotnet/CarbonAware.Tools.WattTimeClient/ConfigurationException.cs
@@ -1,18 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace CarbonAware.Tools.WattTimeClient;
 
-namespace CarbonAware.Tools.WattTimeClient
+/// <summary>
+/// An exception class thrown when the WattTime client is misconfigured.
+/// </summary>
+public class ConfigurationException : Exception
 {
-    /// <summary>
-    /// An exception class thrown when the WattTime client is misconfigured.
-    /// </summary>
-    public class ConfigurationException : Exception
+    public ConfigurationException(string message) : base(message)
     {
-        public ConfigurationException(string message) : base(message)
-        {
-        }
     }
 }
