+++ src/dotnet/CarbonAware.Tools.WattTimeClient/Constants.cs
@@ -1,20 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace CarbonAware.Tools.WattTimeClient;
 
-namespace CarbonAware.Tools.WattTimeClient
+internal class Constants
 {
-    internal class Constants
-    {
-        public const string BalancingAuthorityAbbreviation = "ba";
-        public const string BasicAuthentication = "Basic";
-        public const string BearerAuthentication = "Bearer";
-        public const string StartTime = "starttime";
-        public const string EndTime = "endtime";
-        public const string Latitude = "latitude";
-        public const string Longitude = "longitude";
-        public const string Username = "username";
-    }
+    public const string BalancingAuthorityAbbreviation = "ba";
+    public const string BasicAuthentication = "Basic";
+    public const string BearerAuthentication = "Bearer";
+    public const string StartTime = "starttime";
+    public const string EndTime = "endtime";
+    public const string Latitude = "latitude";
+    public const string Longitude = "longitude";
+    public const string Username = "username";
 }
