+++ src/dotnet/CarbonAware.Tools.WattTimeClient/Paths.cs
@@ -1,17 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace CarbonAware.Tools.WattTimeClient;
 
-namespace CarbonAware.Tools.WattTimeClient
+internal class Paths
 {
-    internal class Paths
-    {
-        public const string Data = "data";
-        public const string Forecast = "forecast";
-        public const string BalancingAuthorityFromLocation = "ba-from-loc";
-        public const string Login = "login";
-        public const string History = "history";
-    }
+    public const string Data = "data";
+    public const string Forecast = "forecast";
+    public const string BalancingAuthorityFromLocation = "ba-from-loc";
+    public const string Login = "login";
+    public const string History = "history";
 }
